GCP/02_Code/01_CloudFunction/GCP/requirements.txt
google-cloud-speech
google-cloud-storage
soundfile
numpy
